further cleaning up after refacoring
