+++ packages/node-wot-servient/src/resource-listeners/basic-resource-listener.ts
@@ -16,6 +16,7 @@
  * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
  * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
+import {ResourceListener,Content} from "node-wot-protocols"
 
 export default class BasicResourceListener implements ResourceListener {
