+++ packages/node-wot-servient/src/resource-listeners/event-resource-listener.ts
